+++ CHANGELOG.md
@@ -5,6 +5,8 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 
 # Unreleased
 
+- [`Fix`] Fix all unhandled errors using ineffassign to find them
+
 # v0.2.0 (6/10/2024)
 
 - [`Breaking`] Some refactoring of names to be proper camel casing