+++ CHANGELOG.md
@@ -11,6 +11,7 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 - Add more functions available from under util
 - Add CoinBatchTransferPayload for sending multiple amounts to multiple addresses
 - [`Fix`] Block APIs will now pull the rest of the transactions for the block automatically
+- [`Fix`] Fix bytecode JSON parsing in transaction parsing
 
 # v0.2.0 (6/10/2024)