+++ CHANGELOG.md
@@ -5,6 +5,8 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 
 # Unreleased
 
+# v0.4.0 (6/17/2024)
+
 - [`Fix`] Fix all unhandled errors using ineffassign to find them
 - [`Doc`] Add examples and documentation in Go doc comments
 - [`Breaking change`] Make SignedTransaction.Hash() output string to be consistent with other representations
@@ -12,7 +14,7 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 - Add CoinBatchTransferPayload for sending multiple amounts to multiple addresses
 - [`Fix`] Block APIs will now pull the rest of the transactions for the block automatically
 - [`Fix`] Fix bytecode JSON parsing in transaction parsing
-- Add ConcClient a concurrent client for many operations
+- Add concurrent APIs, and a transaction handler for a single account
 - Add SimulateTransaction API for single-key authenticated transactions
 
 # v0.2.0 (6/10/2024)